packages/backend/requirements.txt
fastapi
uvicorn[standard]
Pillow
pytest
pytest-asyncio
pytest-watch
requests
